feat!: update create function
Create function now uses snacks.nvim picker module. With this now it is
possible to use same pane for both selecting and creating new notes.
Only problem is there is not a new note selector in the pane, so unless
fuzzy finder does not find anything, it is not possible to create a new
note :D